requirements.txt
connexion == 1.1.15
python_dateutil == 2.6.0
setuptools >= 21.0.0
pymysql
Flask-MySQL == 1.4.0
virtualenv == 16.7.9